+++ CHANGELOG.md
@@ -16,11 +16,11 @@ Categories Used:
 
 **Bullet points in chronological order by PR**
 
-## [Unreleased](https://github.com/ouch-org/ouch/compare/0.8.1...HEAD)
+# For any versions higher than 0.8.1
 
-### Tweaks
+This changelog shouldn't be used anymore.
 
-- Releases: sign assets with cosign instead of GitHub artifact attestations
+This file will be removed soon.
 
 ## [0.8.1](https://github.com/ouch-org/ouch/compare/0.8.0...0.8.1)